# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Multi-tier web data extraction pipeline. Tier 1 performs plain static HTTP fetches.",
        "title": "bextract API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/bextract"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/tier1/fetch": {
            "post": {
                "description": "Performs a plain HTTP GET (or fetches an API endpoint directly)\nusing a rotating User-Agent pool. No JavaScript execution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tier1"
                ],
                "summary": "Tier 1 static HTTP fetch",
                "parameters": [
                    {
                        "description": "Fetch parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tier1handler.FetchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tier1handler.FetchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tier1handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/tier1handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tier2/analyze": {
            "post": {
                "description": "Fetches the URL via Tier 1 then runs the full Tier 2 five-stage pipeline:\nheader analysis, HTML parse, hollow detection, concurrent extraction, and merge.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tier2"
                ],
                "summary": "Tier 2 content detection \u0026 extraction",
                "parameters": [
                    {
                        "description": "Analyze parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tier2handler.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tier2handler.AnalyzeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tier2handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/tier2handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "tier1handler.ErrorResponse": {
            "description": "Error detail returned on failure.",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "dial tcp: connection refused"
                }
            }
        },
        "tier1handler.FetchRequest": {
            "description": "Target URL and optional overrides for the Tier 1 static fetch.",
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "api_endpoint": {
                    "type": "string",
                    "example": "https://api.example.com/data"
                },
                "timeout_ms": {
                    "type": "integer",
                    "example": 5000
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "tier1handler.FetchResponse": {
            "description": "Raw HTTP response metadata and body returned by Tier 1.",
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "example": "\u003chtml\u003e...\u003c/html\u003e"
                },
                "body_size": {
                    "type": "integer",
                    "example": 14823
                },
                "content_type": {
                    "type": "string",
                    "example": "text/html"
                },
                "elapsed_ms": {
                    "type": "integer",
                    "example": 342
                },
                "final_url": {
                    "type": "string",
                    "example": "https://example.com/"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "tier2handler.AnalyzeRequest": {
            "description": "URL and optional configuration for Tier 2 analysis.",
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "api_endpoint": {
                    "type": "string",
                    "example": "https://api.example.com/product/123"
                },
                "extraction_timeout_ms": {
                    "type": "integer",
                    "example": 5000
                },
                "fetch_timeout_ms": {
                    "type": "integer",
                    "example": 10000
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/product/123"
                }
            }
        },
        "tier2handler.AnalyzeResponse": {
            "description": "Tier 2 analysis result including decision, hollow detection, and extracted fields.",
            "type": "object",
            "properties": {
                "decision": {
                    "type": "string",
                    "example": "Done"
                },
                "elapsed_ms": {
                    "type": "integer",
                    "example": 87
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/tier2handler.ExtractedFieldResponse"
                    }
                },
                "hollow_score": {
                    "type": "number",
                    "example": 0.12
                },
                "is_hollow": {
                    "type": "boolean",
                    "example": false
                },
                "is_link_rich": {
                    "type": "boolean",
                    "example": false
                },
                "tech_hints": {
                    "$ref": "#/definitions/tier2handler.TechHintsResponse"
                }
            }
        },
        "tier2handler.ErrorResponse": {
            "description": "Error detail returned on failure.",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "dial tcp: connection refused"
                }
            }
        },
        "tier2handler.ExtractedFieldResponse": {
            "description": "A data field extracted by Tier 2, with provenance metadata.",
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number",
                    "example": 0.95
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "source": {
                    "type": "string",
                    "example": "json-ld"
                },
                "value": {
                    "type": "string",
                    "example": "29.99"
                }
            }
        },
        "tier2handler.TechHintsResponse": {
            "description": "Technology fingerprints detected from HTTP headers.",
            "type": "object",
            "properties": {
                "cf_challenge": {
                    "type": "boolean",
                    "example": false
                },
                "is_cloudflare": {
                    "type": "boolean",
                    "example": false
                },
                "is_json": {
                    "type": "boolean",
                    "example": false
                },
                "is_nextjs": {
                    "type": "boolean",
                    "example": true
                },
                "is_php": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}